Extract Regular Sync Spec Fixture into a separate file
